// src/components/pages/farm-input-product-in-outs/Datatable.type.ts
export enum ApiUrlEnum {
    ALL = '/farm-inputs/product-in-outs/datatable',
    OPNAME = '/farm-inputs/product-in-outs/datatable?type=opname',
    PURCHASE = '/farm-inputs/product-in-outs/datatable?type=purchase',
    SELL = '/farm-inputs/product-in-outs/datatable?type=sell',
    RETURN = '/farm-inputs/product-in-outs/datatable?type=return',
}